revert discordLog
