+++ handlers/commands.js
@@ -3,13 +3,13 @@ const { totalmem } = require("os");
 const discord = require("discord.js");
 const { DateTime, Duration } = require("luxon");
 const log = require("debug")("niles:cmd");
-const { discordLog } = require("./discordLog.js");
 const { i18n } = require("./strings.js");
 const { client } = require("../bot.js");
 const settings = require("../settings.js");
 const helpers = require("./helpers.js");
 const guilds = require("./guilds.js");
 const updaterList = require("./updaterList.js");
+const discordLog = helpers.log;
 const eventType = helpers.eventType;
 const { doHandler } = require("./displayoptions.js");
 const gCalendar = require("@googleapis/calendar").calendar;
