+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node index.js",
-    "windowsDebug": "@powershell -Command $env:DEBUG='*';node index.js"
+    "windowsDebug": "@powershell -Command $env:DEBUG='niles:*';node index.js"
   },
   "author": "Sean Coffey",
   "license": "MIT",
